ai_service/requirements.txt
numpy
opencv-python
scipy
onnxruntime
Pillow
requests
tenacity
torch
transformers
scikit-image
